Included a Regex Exclusion and Updated Example
We need to account for the possibility of the angle bracket being listed as an end to the URL string, so I excluded it from being a part of the match. I also made sure to update the example just to make sure we have a test case that runs with the angle brackets not updating while also showing how it should work to those who read the documentation.